+++ pkg/api/types.go
@@ -133,6 +133,12 @@ func (r *Resource) UnmarshalYAML(value *yaml.Node) error {
 			return fmt.Errorf("%s spec: %w", r.ID(), err)
 		}
 		r.Spec = spec
+	case "IPv6RAAddress":
+		var spec IPv6RAAddressSpec
+		if err := raw.Spec.Decode(&spec); err != nil {
+			return fmt.Errorf("%s spec: %w", r.ID(), err)
+		}
+		r.Spec = spec
 	case "IPv6PrefixDelegation":
 		var spec IPv6PrefixDelegationSpec
 		if err := raw.Spec.Decode(&spec); err != nil {